feat: Make Rust installation persistent
Currently, Rust toolchain is the part of the container image, which
doesn't guarantee that any changes made to it are persistent.

For example, if someone does:

```
icedragon run -- rustup component add llvm-tools-preview
```

Or:

```
icedragon cargo -- install cargo-hack
```

The installed artifacts are wiped out as soon as a new container image
is published on ghcr.io, because icedragon downloads it and replaces
the whole sysroot with the new image contents.

To make artifacts installed with `cargo` or `rustup` persistent,
remove Rust toolchain from the container image and instead, create bind
mounts for `~/.cargo` and `~/.rustup` and bootstrap the toolchain in
a separate step. That makes the lifecycle of the Rust toolchain
independent from the lifecycle of the main image.

Fixes: #10