Build the container image without Dockerfile
The biggest problem with Dockerfiles is that they don't allow extraction
of repetitive parts. Given that we create multiple sysroots and perform
the same actions for each of them, our Dockerfile became very messy.

Avoid that by building the image as a using our containerization and
producing it as a tarball. Then use container engines only to import
that tarball and convert them into an OCI image.