Populate environment with the content of `/etc/environment.d/*` files
`/etc/environment.d` is the standard location for Gentoo to store
system-wide environment variables. Instead of hard coding `PATH`, use
these files to build the environment for containerized processes.

Differentiate between `prepare_env` and `prepare_cross_env`. The former
is going to be used on its own in the next changes.